test changes for clid translation
